# Improve performance and reduce allocations in RouteParameterParser.