@drieseng 2018-11-03
As noted in the PR description, I've started using **ArrayBuilder\<T>** from CoreFX.
If this is considered ok, I'll use it in more classes (eg. **RoutePatternParser**) in future PRs.

**Note:**
See https://github.com/dotnet/corefx/issues/33169 for a discussion on making **ArrayBuilder\<T>** public.