@JamesNK 2018-11-04
This doesn't hurt but parsing routes isn't a bottleneck of ASP.NET apps. It is likely only done once at startup. URL matching and link generation are what happens every request.